+++ components/SimpleGrid/index.js
@@ -27,13 +27,14 @@ export class SimpleGrid extends Component {
             <table className={this.getStyle(this.props.mainClassName)}>
                 {!this.props.hideHeader && <Header
                   externalStyle={this.props.externalStyle}
+                  transformCellValue={this.props.transformCellValue}
                   fields={this.props.fields}
                   toggleColumnVisibility={this.props.toggleColumnVisibility}
                   orderBy={this.props.orderBy}
                   multiSelect={this.props.multiSelect}
                   handleOrder={this.props.handleOrder}
                   isChecked={this.handleIsChecked()}
-                  menuColumn={this.props.menuColumn}
+                  globalMenu={this.props.globalMenu}
                   handleHeaderCheckboxSelect={this.handleHeaderCheckboxSelect}
                 />}
                 <Body
@@ -44,7 +45,7 @@ export class SimpleGrid extends Component {
                   rowsRenderLimit={this.props.rowsRenderLimit}
                   rowsRenderLimitExceedMsg={this.props.rowsRenderLimitExceedMsg}
                   multiSelect={this.props.multiSelect}
-                  menuColumn={this.props.menuColumn}
+                  globalMenu={this.props.globalMenu}
                   transformCellValue={this.props.transformCellValue}
                   handleCheckboxSelect={this.props.handleCheckboxSelect}
                   handleCellClick={this.props.handleCellClick}
@@ -65,7 +66,7 @@ SimpleGrid.propTypes = {
     hideHeader: PropTypes.bool,
     handleOrder: PropTypes.func,
     rowsChecked: PropTypes.array,
-    menuColumn: PropTypes.bool,
+    globalMenu: PropTypes.bool,
     transformCellValue: PropTypes.func,
     toggleColumnVisibility: PropTypes.func,
     mainClassName: PropTypes.string,