+++ containers/LoginForm/reducer.js
@@ -1,27 +1,45 @@
 import Immutable from 'immutable';
-import { LOGIN, SET_INPUT_VALUE, VALIDATE_FORM, COOKIE_CHECK, LOGOUT } from './actionTypes';
-import { getInputs, inputs as inputsConfig } from './config';
+import { LOGIN, BIO_LOGIN, SET_INPUT_VALUE, VALIDATE_FORM, COOKIE_CHECK, LOGOUT } from './actionTypes';
+import { inputs as inputsConfig, loginSteps } from './config';
 import { Validator } from './../../utils/validator';
 
 const validator = new Validator(inputsConfig);
 
+const updateLoginStep = (state, step) => {
+    let loginStep = loginSteps[step];
+    let currentInputs = state.getIn(['loginForm', 'inputs']);
+    let newInputs = Immutable.Map();
+
+    Object.keys(loginStep.inputs).forEach(input => {
+        let inputValue = state.getIn(['loginForm', 'inputs', input, 'value']);
+        // if the form already has this input, perserve its value only and reset its cofig
+        let newInput = currentInputs.has(input) ? Object.assign({}, loginStep.inputs[input], { value: inputValue }) : loginStep.inputs[input];
+        newInputs = newInputs.set(input, Immutable.fromJS(newInput));
+    });
+
+    state = state.setIn(['loginForm', 'inputs'], newInputs);
+
+    loginStep.disabledFields.forEach(field => {
+        state = state.setIn(['loginForm', 'inputs', field, 'disabled'], true);
+    });
+
+    return state.setIn(['loginForm', 'buttonLabel'], loginStep.buttonLabel)
+                .setIn(['loginForm', 'title'], loginStep.title)
+                .set('formError', '')
+                .set('loginType', step);
+};
+
 const defaultLoginState = Immutable.fromJS({
     authenticated: false,
-    shouldChangePassword: false,
-    loginFailedAttepmts: 0,
     cookieChecked: false,
-    loginForm: {
-        inputs: getInputs(['username']),
-        formError: '',
-        shouldSubmit: false,
-        invalidField: '',
-        title: 'Login',
-        buttonLabel: 'Next'
-    },
+    loginForm: loginSteps['initial'],
+    loginType: '',
+    formError: '',
+    shouldSubmit: false,
     loginData: {}
 });
 
-const updateLoginData = (state) => {
+const getLoginData = (state) => {
     let inputs = state.getIn(['loginForm', 'inputs']);
     let currentLoginData = state.get('loginData');
 
@@ -42,46 +60,41 @@ export const login = (state = defaultLoginState, action) => {
         case LOGIN:
             if (action.methodRequestState === 'finished') {
                 state = state.setIn(['loginForm', 'shouldSubmit'], false);
-                // show password input and change title
-                if (action.error && action.error.type === 'policy.param.password') {
-                    return state.setIn(['loginForm', 'inputs', 'password'], Immutable.fromJS(getInputs(['password']).password))
-                                .setIn(['loginForm', 'inputs', 'username', 'disabled'], true)
-                                .setIn(['loginForm', 'title'], Immutable.fromJS('Login with password'))
-                                .setIn(['loginForm', 'buttonLabel'], 'Login');
-                } else if (action.error && action.error.type === 'policy.param.newPassword') {
-                    // take only the username input from the current state and merge it with the new inputs for this step
-                    let newInputs = state.getIn(['loginForm', 'inputs']).take(1).merge(getInputs(['newPassword', 'confirmPassword']));
 
-                    return state.setIn(['loginForm', 'title'], 'Password change required')
-                                .setIn(['loginForm', 'inputs'], newInputs)
-                                .setIn(['loginForm', 'buttonLabel'], 'Change');
-                } else if (action.error) {
-                    return state.setIn(['loginForm', 'formError'], action.error.message);
+                if (action.error) {
+                    let err = action.error.type.split('.');
+                    let type = err[err.length - 1];
+
+                    return loginSteps[type] ? updateLoginStep(state, type) : state.set('formError', action.error.message);
                 } else if (action.result) {
                     return state.set('authenticated', true)
-                                .setIn(['loginForm', 'formError'], '')
                                 .set('cookieChecked', true)
-                                .setIn(['loginForm', 'formError'], '');
+                                .set('formError', '');
                 }
             }
             return state;
+        case BIO_LOGIN:
+            return state;
         case SET_INPUT_VALUE:
-            let { input, value } = action;
-            return state.setIn(['loginForm', 'inputs', input, 'value'], value);
+            return state.setIn(['loginForm', 'inputs', action.input, 'value'], action.value);
         case VALIDATE_FORM:
-            let inputs = state.getIn(['loginForm', 'inputs']);
-            let loginData = state.get('loginData');
-
-            validationResult = validator.validateAll(inputs);
+            validationResult = validator.validateAll(state.getIn(['loginForm', 'inputs']));
 
             if (validationResult.isValid) {
-                loginData = updateLoginData(state);
+                let prevInvalidField = state.getIn(['loginForm', 'inputs']).find(input => input.get('error'));
+
+                if (prevInvalidField) {
+                    state = state.setIn(['loginForm', 'inputs', prevInvalidField.get('name'), 'error'], '');
+                }
+
+                state = state.set('loginData', getLoginData(state));
+            } else {
+                state = state.setIn(['loginForm', 'inputs', validationResult.invalidField, 'error'], validationResult.error)
+                              .set('formError', '');
             }
 
-            return state.set('loginData', loginData)
-                        .setIn(['loginForm', 'shouldSubmit'], validationResult.isValid)
-                        .setIn(['loginForm', 'formError'], validationResult.error)
-                        .setIn(['loginForm', 'invalidField'], validationResult.invalidField);
+            return state.setIn(['loginForm', 'shouldSubmit'], validationResult.isValid);
+
         case COOKIE_CHECK:
             if (action.methodRequestState === 'finished') {
                 if (action.error) {